+++ app/helpers/repositories_helper.rb
@@ -1,5 +1,0 @@
-module RepositoriesHelper
-  def can_star_repository?(repository)
-    !repository.starred_by? current_user
-  end
-end